+++ app/controllers/logins_controller.rb
@@ -1,8 +1,8 @@
+# frozen_string_literal: true
 class LoginsController < ApplicationController
   include JwtHelper
 
   def create
-    debugger
     user = UserBlock::User.find_by(email: params[:email])
     if user && user.authenticate(params[:password])
       payload = { user_id: user.id }
@@ -12,4 +12,5 @@ class LoginsController < ApplicationController
       render json: { error: 'Invalid email or password' }, status: :unauthorized
     end
   end
+
 end
