+++ config/routes.rb
@@ -5,9 +5,16 @@ Rails.application.routes.draw do
   # Defines the root path route ("/")
   # root "articles#index"  resources :users
 
-  resources :admins
+  namespace :user_block do
+    resources :admins
+    resources :users
+  end
+  post 'sign_up', to: 'user_block/users#create'
+
   resources :allocation_drafts
 
   post "import_allocation", to: "allocation_drafts#import"
   post 'login', to: 'logins#create'
+
+  get 'types', to: 'public_apis#types'
 end
